moviebookingapp_frontend/src/app/services/hall.service.ts
import {
  HttpHeaders,
  HttpClient,
  HttpErrorResponse,
} from '@angular/common/http';
import { Injectable } from '@angular/core';
import { Observable, catchError, throwError } from 'rxjs';
import { URLs } from '../_shared/urls';
import { Movie } from '../models/movie';
import { Hall } from '../models/hall';

@Injectable({
  providedIn: 'root',
})
export class HallService {
  httpOptions = {
    headers: new HttpHeaders({
      'Content-Type': 'application/json',
      'Access-Control-Allow-Origin': '*',
    }),
  };
  constructor(private httpClient: HttpClient) {}

  getHallById(id: string): Observable<Hall> {
    return this.httpClient
      .get<Hall>(URLs.HALL_API_BASE_URL + '/search/' + id, this.httpOptions)
      .pipe(catchError(this.handleError));
  }

  handleError(eResponse: HttpErrorResponse) {
    if (eResponse.error instanceof ErrorEvent) {
      console.log('Client Side Error =' + eResponse.error.message);
      console.log('Status Code=' + eResponse.status);
    } else {
      console.log('Server Side Error =' + eResponse.error.message);
      console.log('Status Code=' + eResponse.status);
    }
    return throwError(eResponse.error.message);
  }
}
